fix(geiger_toolbox): add offering to todo, watchTodoOfferingStatus ha…

+++ apps/geiger_toolbox/lib/src/features/threat_assessment/presentation/clear_data/clear_data_button.dart
@@ -1,10 +1,13 @@
 import 'package:core_ui/molecules/buttons/app_button.dart';
+import 'package:core_ui/molecules/buttons/app_icon_button.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:geiger_toolbox/src/common_widgets/alert_dialogs.dart';
 import 'package:geiger_toolbox/src/extensions/async_value_extension.dart';
 import 'package:geiger_toolbox/src/features/threat_assessment/presentation/clear_data/clear_data_controller.dart';
 
 import 'package:geiger_toolbox/src/localization/string_hardcoded.dart';
+import 'package:go_router/go_router.dart';
 
 class ClearDataButton extends ConsumerWidget {
   const ClearDataButton({super.key});
@@ -14,17 +17,27 @@ class ClearDataButton extends ConsumerWidget {
     ref.listen(cleanDataControllerProvider,
         (_, nxt) => nxt.showAlertDialogOnError(context: context));
     final state = ref.watch(cleanDataControllerProvider);
-
-    return AppButton.delete(
+    final colors = Theme.of(context).colorScheme;
+    return AppIconButton.filled(
       context: context,
-      label: state.isLoading ? 'deleting' : "Clear Cache".hardcoded,
+      backgroundColor: colors.error,
+      iconData: state.isLoading ? Icons.timer : Icons.delete_forever,
       onPressed: state.isLoading
           ? null
           : () async {
-              await ref
-                  .read(cleanDataControllerProvider.notifier)
-                  .deleteNewsData();
-              await ref.read(cleanDataControllerProvider.notifier).deleteTask();
+              showAlertDialog(
+                  context: context,
+                  title: "Alert".hardcoded,
+                  content:
+                      "Are you sure you want to delete clean? \n Note: This action can't be undone.",
+                  cancelActionText: 'Cancel'.hardcoded,
+                  trigger: () {
+                    ref
+                        .read(cleanDataControllerProvider.notifier)
+                        .deleteNewsData();
+                    //pop alert
+                    context.pop(true);
+                  });
             },
     );
   }
